Authorization through ACLs in gatekeeper
Gatekeeper requires a new parameter called `acls` which describes which
app can reach which paths on which endpoints.  There's an example:

  - app: fred
    endpoints:
      - endpoint: wilma
        paths: ['/', '/dinner/:date']
      - endpoint: betty
        paths: ['/visits']

The endpoints are taken from the X-Route header which is formatted as
follows:

  X-Route: endpoint=wilma,betty

The gatekeeper enforces these ACLs using the user part of the
authentication and matching that as `app` to the above ACL description.
When the request does not have an X-Route header or tries an
endpoint/path combination not present in the ACls, it will respond with
a 403 Forbidden response.