OAuth2 Client Credentials support
A service endpoint can not be configured to use OAuth2 Client
Credentials as authorization by adding proxyOptions.  Here's an example:

  SomeBackend:
    url: https://some.backend/
    proxyOptions:
      oauth2:
        clientCredentials:
          tokenEndpoint:
            url: https://some-auth-server/token
            params:
              grant_type: client_credentials
              client_id: fred
              client_secret: wilma

Tokens acquired from the token endpoint will be cached when the token
endpoint includes an "expires_in" property.

When one of the endpoints in a request is authenticated through OAuth2
but the post-token endpoint of that request responds badly, the entire
request should fail and an error message is logged.

The way of authentication is tested through a mock oauth authorization
service.  This service only facilitates client_credentials grants, as
does the client part, and returns a bearer token with the following
format:

  good-token/{tstamp}

The placeholder tstamp represents an ISO timestamp.  The token will
expire in 5 seconds, the timestamp is used by the test-backend to test
the validity of the token.

This implementation is based on RFC 6749, specifically on sections 3.2
and 4.4.