Timing attack mitigation
Given application passwords and salts are always 32 characters long (as
already enforced by the included "credentials" CLI utility and the
external "configurator" application) always "do the work" when
authenticating an application.

In this changeset a password hash is always calculated on a 65 character
string (32 characters password, a dash and 32 characters salt) and
crypto.timingSafeEqual is used.  If a attacker provides an unknown user
name, a dummy hash and salt will be provided to do work on and the
provided password will be padded or truncated to 32 characters to ensure
the amount of work is the same as for a known user.

Note: this change is backward compatible because both passwords and
salts distributed are 32 characters long.

Also note: this implementation assumes no magic by the runtime JIT or
caching by the hashing function.