Fix authentication checks for ACLs containing no paths.
If the ACL for a particular app -> endpoint pair contained an empty
collection of paths, the app would be granted access to all paths in
the collection.

The problem was that matching was done by a compiled regex:

```
new RegExp(paths.map(path => pathToRegexp(path).source).join('|'))
```

Which for an empty `paths` collection results in a RegExp(""), which
matches all strings.

In addition, if a request without an x-routes header was received, all
endpoints in the ACL, including the ones without paths, would be used.

This fixes the issues by:
  - explicitly testing for an empty collection and returning no
     matcher in that case
  - not prefilling x-routes for endpoints that have no matcher