DEV: use local copy of express-gateway dependency
This introduces our fork of express gateway as a submodule in
vendor/express-gateway and sets up the tests and docker files to use
it.

Note that this currently slows down the tests because node modules are
imported on every test run - this ensures you're always running tests
against the current version of vendor/express-gateway.

This commit should be reverted/removed later.