Adds healthcheck (#137)
This PR adds a health check to the Dockerfile. It is purely meant to check whether the container has successfully started and is capable of accepting HTTP requests. This allows for zero downtime rolling deploys using Docker Swarm.

More “in depth checking” to make sure the container is actually capable of functioning as a proxy will be done by external monitoring tools.