Add validation tests for all paths in OOAPI
Adds integration test code that will request each specified path in
the ooapi spec to ensure that the validator code can actually validate
the request (and response, for v4).

Note that the ooapi v5 spec validates only that the response is a json
object, which is the reason we can use the same test data for v4 and
v5 tests.  The data that is only used for v5 tests is just an empty
object `{}` for the same reason.

Also not that some v4 paths are not tested since the example data for
those tests (from the v4 documentation) is not actually valid.

This change also updates the v4 spec used by the gateway, and adds 2
new paths to the v4 gateway configuration:

 - '/buildings/:buildingId/rooms'
 - '/organizations/:organizationId/offerings'

You can now also generate up-to-date ooapiv4 specs by updating the
`ooapi-specification` submodule and running `make ooapiv4.json`.